requirements.txt
numpy
easyocr
streamlit
pillow
PyMuPDF
matplotlib
shapely
langdetect 
googletrans==4.0.0-rc1